# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "enum": [
          "submitted",
          "under_human_review",
          "withdrawn",
          "closed"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
...
      "CreditApplicationListResponse": {
        "properties": {
          "applications": {
            "items": {
              "$ref": "#/components/schemas/CreditApplicationView"
            },
            "title": "Applications",
            "type": "array"
          }
        },
        "required": [
          "applications"
        ],
        "title": "CreditApplicationListResponse",
        "type": "object"
      },
      "CreditApplicationMoveRequest": {
        "additionalProperties": false,
        "description": "Take an intake into human review, or close it; the version it was read at guards against a concurrent move.",
        "properties": {
          "expected_version": {
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "CreditApplicationMoveRequest",
        "type": "object"
      },
      "CreditApplicationView": {
        "description": "An intake recorded for human review. It is never a lending decision; the policy behind it is synthetic.",
        "properties": {
          "application_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
            "title": "Application Id",
            "type": "string"
          },
          "assessment_ref": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Ref"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
            "title": "Customer Id",
            "type": "string"
          },
          "declared_monthly_income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin_conversation_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Conversation Id"
          },
          "product_code": {
            "pattern": "^[A-Z][A-Z0-9_-]{2,31}$",
            "title": "Product Code",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "requested_amount": {
            "$ref": "#/components/schemas/Money"
          },
          "requested_term_months": {
            "title": "Requested Term Months",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "status_history": {
            "items": {
              "$ref": "#/components/schemas/ApplicationStatusChange"
            },
            "title": "Status History",
            "type": "array"
          },
          "synthetic_policy": {
            "title": "Synthetic Policy",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "description": "Send it back as `expected_version` to move the intake (review, close).",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "application_id",
          "customer_id",
          "product_code",
          "requested_amount",
          "requested_term_months",
          "purpose",
          "declared_monthly_income",
          "assessment_ref",
          "status",
          "version",
          "created_at",
          "updated_at",
          "status_history",
          "origin_conversation_id",
          "synthetic_policy"
        ],
        "title": "CreditApplicationView",
        "type": "object"
      },
...
      "EscalationReasonCode": {
        "enum": [
          "amount_above_auto_limit",
          "repeat_complainer",
          "legal_or_regulator_mention",
          "distress_signal",
          "human_requested",
          "clarification_exhausted",
          "tool_failure",
          "verification_mismatch",
          "risk_tier_high",
          "sla_breached",
          "unsupported_needs_human",
          "other",
          "card_unblock_requested",
          "card_replacement_requested",
          "credit_review_required",
          "eligibility_contested"
        ],
        "title": "EscalationReasonCode",
        "type": "string"
      },
...
      "HandoffListResponse": {
        "properties": {
          "handoffs": {
            "items": {
              "$ref": "#/components/schemas/HandoffView"
            },
            "title": "Handoffs",
            "type": "array"
          }
        },
        "required": [
          "handoffs"
        ],
        "title": "HandoffListResponse",
        "type": "object"
      },
...
      "HandoffStatus": {
        "enum": [
          "open",
          "claimed",
          "resolved"
        ],
        "title": "HandoffStatus",
        "type": "string"
      },
...
      "Language": {
        "enum": [
          "es",
          "pt",
          "en"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Priority",
        "type": "string"
      },
      "ProblemDetails": {
        "description": "RFC 9457 problem details. `type` is a stable URI per error; `errors` lists invalid fields.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "properties": {
                "loc": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": {
                  "type": "string"
                }
              },
              "required": [
                "loc",
                "type"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "ResolveHandoffRequest": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/HandoffOutcomeCode"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "ResolveHandoffRequest",
        "type": "object"
      },
...
      "WorkflowId": {
        "description": "The supported workflows (section 1 of CLAUDE.md). ``WorkflowRef.id`` stays a pattern-constrained string in\nevery contract; membership in this registry is checked by the application, not by the schemas.",
        "enum": [
          "account_inquiry",
          "card_support",
          "dispute",
          "credit"
        ],
        "title": "WorkflowId",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "The double-submit token from `GET /v1/auth/csrf`, echoed on every state-changing request.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "Opaque session token (HttpOnly). The development cookie is named `session`.",
        "in": "cookie",
        "name": "__Host-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "bank-agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent/credit-applications": {
      "get": {
        "description": "Reviewable intakes (submitted, under human review) and any a handoff references, newest first.\n\nRead only, and never a lending decision.",
        "operationId": "agent_list_credit_applications",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationStatus"
                  },
                  "maxItems": 4,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Credit Applications",
        "tags": [
          "agent"
        ],
        "x-csrf": false,
        "x-rate-limit": "read",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/credit-applications/{application_id}": {
      "get": {
        "operationId": "agent_get_credit_application",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditApplicationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Credit Application",
        "tags": [
          "agent"
        ],
        "x-csrf": false,
        "x-rate-limit": "read",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/credit-applications/{application_id}/close": {
      "post": {
        "description": "Close an intake under human review. Audited; there is no approved or declined status.",
        "operationId": "agent_close_credit_application",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditApplicationMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditApplicationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Close Credit Application",
        "tags": [
          "agent"
        ],
        "x-csrf": true,
        "x-rate-limit": "write",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/credit-applications/{application_id}/review": {
      "post": {
        "description": "Take a submitted intake into human review. Audited; never a lending decision.",
        "operationId": "agent_review_credit_application",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditApplicationMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditApplicationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Review Credit Application",
        "tags": [
          "agent"
        ],
        "x-csrf": true,
        "x-rate-limit": "write",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/handoffs": {
      "get": {
        "description": "The inbox, soonest SLA first, filtered by workflow, priority, reason, language, status, and SLA due.",
        "operationId": "agent_list_handoffs",
        "parameters": [
          {
            "in": "query",
            "name": "workflow",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowId"
                  },
                  "maxItems": 4,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Priority"
                  },
                  "maxItems": 4,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/EscalationReasonCode"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  },
                  "maxItems": 3,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/HandoffStatus"
                  },
                  "maxItems": 3,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sla_due_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sla Due Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List Handoffs",
        "tags": [
          "agent"
        ],
        "x-csrf": false,
        "x-rate-limit": "read",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/handoffs/{handoff_id}": {
      "get": {
        "operationId": "agent_get_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Handoff Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get Handoff",
        "tags": [
          "agent"
        ],
        "x-csrf": false,
        "x-rate-limit": "read",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/handoffs/{handoff_id}/claim": {
      "post": {
        "description": "Claim an open handoff for the signed-in agent.",
        "operationId": "agent_claim_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Handoff Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Claim Handoff",
        "tags": [
          "agent"
        ],
        "x-csrf": true,
        "x-rate-limit": "write",
        "x-roles": [
          "agent"
        ]
      }
    },
    "/v1/agent/handoffs/{handoff_id}/resolve": {
      "post": {
        "description": "Resolve a handoff this agent claimed, with an outcome code and a note.",
        "operationId": "agent_resolve_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
              "title": "Handoff Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveHandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error as problem details"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Resolve Handoff",
        "tags": [
          "agent"
        ],
        "x-csrf": true,
        "x-rate-limit": "write",
        "x-roles": [
          "agent"
        ]
      }
    },
...
  }
}